I seriously debated to write this reply because there is an opinion to instantly deny this that is quite valid. Your note history means that in my opinion you should not get on the team at this time.  I say at this time. And know that my post will upset some people. The reason I say at this time is because I do think people grow up and change. There are people who were at thin ice in cm and that did change.  Post Scriptum: The more I look at staff stuff the less I like. The path you have taken/are taking might have ruled you out from being a cm dev forever.  I also say it because I am trying to do my best to be constructive. And to tell you what it takes for me at least before you app will be considered.  I want to avoid posting in public what action is taken against players. So hypotheticly if you have any bans appeal them. Hypotheticly if you get a lot of notes avoid doing so for a bit. I cannot say how much time is enough time. But hypotheticly posting an app within a week of getting a new note is certainly not a good idea. Keep in mind that trying to game this goes against any good faith your attempt to do this generates.  I also want to state that this is my personal opinion. I have not had administrative dealings with you in the past. And that we do take into account the opinion of the rest of staff and community if needed.  Edit: Since I have seen the chat about this and there is 0 chance of this being accepted at this time I am not going to wait for someone to come in and yell denied. At this time it wont happen. If you improve I hope I might have to debate an app one day where you are a long time without notes and issues. And once that happens I am perfectly willing to take the app seriously.
I seriously debated to write this reply because there is an opinion to instantly deny this that is quite valid.
Your note history means that in my opinion you should not get on the team at this time.

I say at this time. And know that my post will upset some people.
The reason I say at this time is because I do think people grow up and change.
There are people who were at thin ice in cm and that did change.

Post Scriptum:
The more I look at staff stuff the less I like.
The path you have taken/are taking might have ruled you out from being a cm dev forever.

I also say it because I am trying to do my best to be constructive.
And to tell you what it takes for me at least before you app will be considered.

I want to avoid posting in public what action is taken against players.
So hypotheticly if you have any bans appeal them.
Hypotheticly if you get a lot of notes avoid doing so for a bit.
I cannot say how much time is enough time.
But hypotheticly posting an app within a week of getting a new note is certainly not a good idea.
Keep in mind that trying to game this goes against any good faith your attempt to do this generates.

I also want to state that this is my personal opinion.
I have not had administrative dealings with you in the past.
And that we do take into account the opinion of the rest of staff and community if needed.

Edit: Since I have seen the chat about this and there is 0 chance of this being accepted at this time I am not going to wait for someone to come in and yell denied.
At this time it wont happen. If you improve I hope I might have to debate an app one day where you are a long time without notes and issues. And once that happens I am perfectly willing to take the app seriously.